requirements.txt
norfair[video]
numpy
Pillow
matplotlib==3.6.1
pandas==1.5.0
PyYAML==6.0
tqdm==4.64.1
seaborn==0.12.0
ipython==8.5.0
ultralytics


# SoccerNet CALF notebook/toolchain extras
torch
torchvision
torchaudio
SoccerNet
scikit-video
tensorflow==2.3.0
imutils
opencv-python==3.4.11.41
moviepy
scikit-learn
ffmpy

# Utility tooling
colorama==0.4.5
requests==2.28.1


